fix(xmem): moved `*mem_to_*mem()` to each `*.c`.
